LPS-109668 | LPS-110342 Improve QuickActionKebab component
