--- modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/src/main/resources/META-INF/resources/js/shared/components/quick-action-kebab/DropDownItem.es.js
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- */
-
-import ClayDropDown from '@clayui/drop-down';
-import React from 'react';
-
-const DropDownItem = ({action, title}) => {
-	return (
-		<ClayDropDown.ItemList>
-			<li>
-				<button
-					className="dropdown-item"
-					data-testid="dropDownItem"
-					onClick={action}
-				>
-					{title}
-				</button>
-			</li>
-		</ClayDropDown.ItemList>
-	);
-};
-
-export {DropDownItem};
